Paula cleanup, State machine cleanup

+++ Emulator/Paula/Audio/StateMachineTypes.h
@@ -17,7 +17,7 @@
 
 typedef struct
 {
-    i8 state;
+    isize state;
     bool dma;
     u16 audlenLatch;
     u16 audlen;
